refactor: Move HitomiError to error.ts
